Added model parsing tests from fixtures

+++ src/main/java/postageapp/models/Message.java
@@ -12,7 +12,7 @@ import java.util.Map;
  */
 public class Message extends PostageModel {
     private String uid, template;
-    private int transmissionsTotal, transmissionsFailed, transmissionsCompleted;
+    private double transmissionsTotal, transmissionsFailed, transmissionsCompleted;
     private Date createdAt, willPurgeAt;
 
     public Message(String uid, Map<String, ?> json) {
@@ -20,9 +20,9 @@ public class Message extends PostageModel {
 
         this.uid = uid;
         this.template = (String) json.get("template");
-        this.transmissionsTotal = (Integer) json.get("transmissions_total");
-        this.transmissionsCompleted = (Integer) json.get("transmissions_completed");
-        this.transmissionsFailed = (Integer) json.get("transmissions_failed");
+        this.transmissionsTotal = (Double) json.get("transmissions_total");
+        this.transmissionsCompleted = (Double) json.get("transmissions_completed");
+        this.transmissionsFailed = (Double) json.get("transmissions_failed");
         this.createdAt = this.dateFromString((String) json.get("created_at"));
         this.willPurgeAt = this.dateFromString((String) json.get("will_purge_at"));
     }
@@ -35,15 +35,15 @@ public class Message extends PostageModel {
         return template;
     }
 
-    public int getTransmissionsCompleted() {
+    public double getTransmissionsCompleted() {
         return transmissionsCompleted;
     }
 
-    public int getTransmissionsFailed() {
+    public double getTransmissionsFailed() {
         return transmissionsFailed;
     }
 
-    public int getTransmissionsTotal() {
+    public double getTransmissionsTotal() {
         return transmissionsTotal;
     }
